requirements.txt
tldextract
python-whois
dnspython